Initial commit of train_whistle, a Factorio mod.
Have you been run down by an errant train that came from off-screen while you weren't looking?
Of course you have! Everyone has. It sucks, and you die.

The goal here is to write lookahead code so that trains toot if you're near where they'll be in the
next few seconds, and send a long blast if you're actually on the tracks.