fix error when factorio tries to load the zip file.
The error thrown by Factorio when it loads a bad mod zip is inaccurate. It wants the zipfile to
contain a reference folder of the same name as the zipfile, which contains the mod's files. It
claims not to want that internal reference folder, though.

Should have believed the docs from the start, I guess.